feat: auth 도메인과 user 도메인 분리 작업
