+++ src/main/java/com/devcourse/ReviewRanger/auth/domain/UserPrincipal.java
@@ -1,4 +1,4 @@
-package com.devcourse.ReviewRanger.user.domain;
+package com.devcourse.ReviewRanger.auth.domain;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -7,6 +7,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import com.devcourse.ReviewRanger.user.domain.User;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 public class UserPrincipal implements UserDetails {
