+++ src/main/java/com/devcourse/ReviewRanger/user/api/UserController.java
@@ -7,23 +7,16 @@ import javax.validation.Valid;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.devcourse.ReviewRanger.auth.domain.UserPrincipal;
 import com.devcourse.ReviewRanger.common.response.RangerResponse;
 import com.devcourse.ReviewRanger.user.application.UserService;
-import com.devcourse.ReviewRanger.user.domain.UserPrincipal;
 import com.devcourse.ReviewRanger.user.dto.GetUserResponse;
-import com.devcourse.ReviewRanger.user.dto.JoinRequest;
-import com.devcourse.ReviewRanger.user.dto.LoginRequest;
-import com.devcourse.ReviewRanger.user.dto.LoginResponse;
 import com.devcourse.ReviewRanger.user.dto.UpdateNameRequest;
 import com.devcourse.ReviewRanger.user.dto.UpdatePasswordRequest;
 import com.devcourse.ReviewRanger.user.dto.UserInfoResponse;
-import com.devcourse.ReviewRanger.user.dto.ValidateEmailRequest;
-import com.devcourse.ReviewRanger.user.dto.ValidateNameRequest;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -37,68 +30,6 @@ public class UserController {
 
 	public UserController(UserService userService) {
 		this.userService = userService;
-	}
-
-	@Tag(name = "user")
-	@Operation(summary = "회원가입", description = "사용자가 회원가입을 요청하는 API", responses = {
-		@ApiResponse(responseCode = "200", description = "회원가입 성공"),
-		@ApiResponse(responseCode = "409", description = "이미 가입된 아이디를 입력한 경우"),
-		@ApiResponse(responseCode = "409", description = "이미 가입된 이메일을 입력한 경우")
-	})
-	@PostMapping("/sign-up")
-	public RangerResponse<Void> join(@RequestBody @Valid JoinRequest request) {
-		Boolean joinSuccess = userService.join(request);
-
-		return RangerResponse.ok(joinSuccess);
-	}
-
-	@Tag(name = "user")
-	@Operation(summary = "로그인", description = "사용자가 로그인을 요청하는 API", responses = {
-		@ApiResponse(responseCode = "200", description = "로그인 성공"),
-		@ApiResponse(responseCode = "422", description = "잘못된 로그인 정보를 입력한 경우"),
-	})
-	@PostMapping("/login")
-	public RangerResponse<LoginResponse> login(@RequestBody @Valid LoginRequest request) {
-		LoginResponse loginResponse = userService.login(request);
-
-		return RangerResponse.ok(loginResponse);
-	}
-
-	@Tag(name = "user")
-	@Operation(summary = "[토큰] 로그아웃", description = "[토큰] 사용자가 로그아웃을 요청하는 API", responses = {
-		@ApiResponse(responseCode = "200", description = "로그아웃 성공"),
-		@ApiResponse(responseCode = "401", description = "토큰을 넣지 않은 경우"),
-	})
-	@PostMapping("/members/logout")
-	public RangerResponse<Void> logout(
-		@RequestHeader("Authorization") String accessToken
-	) {
-		userService.logout(accessToken);
-		return RangerResponse.noData();
-	}
-
-	@Tag(name = "user")
-	@Operation(summary = "이름 중복검사", description = "이름 중복 검사 API", responses = {
-		@ApiResponse(responseCode = "200", description = "이름이 중복되지 않음"),
-		@ApiResponse(responseCode = "409", description = "이미 가입된 이름을 입력한 경우")
-	})
-	@PostMapping("/members/check-name")
-	public RangerResponse<Void> checkname(@RequestBody @Valid ValidateNameRequest request) {
-		boolean notExistName = userService.isNotExistName(request.name());
-
-		return RangerResponse.ok(notExistName);
-	}
-
-	@Tag(name = "user")
-	@Operation(summary = "이메일 중복검사", description = "이메일 중복 검사 API", responses = {
-		@ApiResponse(responseCode = "200", description = "이메일이 중복되지 않음"),
-		@ApiResponse(responseCode = "409", description = "이미 가입된 이메일을 입력한 경우")
-	})
-	@PostMapping("/members/check-email")
-	public RangerResponse<Void> checkEmail(@RequestBody @Valid ValidateEmailRequest request) {
-		boolean notExistEmail = userService.isNotExistEmail(request.email());
-
-		return RangerResponse.ok(notExistEmail);
 	}
 
 	@Tag(name = "user")
