+++ src/main/java/com/devcourse/ReviewRanger/user/dto/UpdatePasswordRequest.java
@@ -7,6 +7,7 @@ import javax.validation.constraints.Pattern;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+@Schema(description = "사용자 비밀번호 변경 요청 DTO")
 public record UpdatePasswordRequest(
 	@Schema(description = "변경 할 비밀번호")
 	@NotBlank(message = "비밀번호는 빈값 일 수 없습니다.")
